no implicit ValueReader for Int128, making a Long for now

+++ src/main/scala/co/topl/settings/AppSettings.scala
@@ -71,7 +71,7 @@ case class NetworkSettings(
 
 case class ForgingSettings(
   blockGenerationDelay: FiniteDuration,
-  minTransactionFee:    Int128,
+  minTransactionFee:    Long,
   protocolVersions:     List[ProtocolSettings],
   forgeOnStartup:       Boolean,
   privateTestnet:       Option[PrivateTestnetSettings]
